#4316 move team logic to util
