+++ src/frontend/src/utils/rules.utils.ts
@@ -90,3 +90,117 @@ export const getAncestorIds = (ruleId: string, allRules: Rule[]): string[] => {
  */
 export const compareRuleCodes = (a: Rule, b: Rule): number =>
   a.ruleCode.localeCompare(b.ruleCode, undefined, { numeric: true });
+
+export interface TeamProject {
+  projectId: string;
+  projectName: string;
+  rules: Rule[];
+}
+
+export interface TeamRules {
+  teamId: string;
+  teamName: string;
+  projects: TeamProject[];
+  unassignedRules: Rule[];
+}
+
+/**
+ * Groups ruleset's rules by team and project for team view. First groups by team
+ * assigned to, then by projects assign to within that team, or unassigned to any
+ * project. Finally groups by rules not assigned to any team.
+ * Note that rules can be on more than one team and in more than one project for that team.
+ */
+export const getTeamOrganization = (allRules: Rule[]): { teamRules: TeamRules[]; unassignedToTeam: Rule[] } => {
+  const teamMap = new Map<string, TeamRules>();
+  const teamProjectMaps = new Map<string, Map<string, TeamProject>>();
+  const unassignedToTeam: Rule[] = [];
+
+  // iterate through all rules and organize by team and project
+  allRules.forEach((rule) => {
+    // rule with no team belongs to the top-level "unassigned to team" bucket.
+    if (!rule.teams || rule.teams.length === 0) {
+      unassignedToTeam.push(rule);
+      return;
+    }
+
+    // otherwise place the rule under each team it belongs to
+    rule.teams.forEach((team) => {
+      // create the team's buckets the first time we see it
+      if (!teamMap.has(team.teamId)) {
+        teamMap.set(team.teamId, {
+          teamId: team.teamId,
+          teamName: team.teamName,
+          projects: [],
+          unassignedRules: []
+        });
+        teamProjectMaps.set(team.teamId, new Map());
+      }
+
+      const teamRules = teamMap.get(team.teamId)!;
+      const projectMap = teamProjectMaps.get(team.teamId)!;
+
+      // Keep only the rule's projects that belong to this team. Since rules can be on multiple teams,
+      // this ensures projects don't show up under teams they are not a part of
+      const projectsForTeam = rule.projects?.filter((project) => project.teamIds.includes(team.teamId)) ?? [];
+
+      // Add the rule under each of its projects, or unassigned to project, for this team
+      if (projectsForTeam.length > 0) {
+        projectsForTeam.forEach((project) => {
+          if (!projectMap.has(project.projectId)) {
+            projectMap.set(project.projectId, {
+              projectId: project.projectId,
+              projectName: project.projectName,
+              rules: []
+            });
+          }
+          projectMap.get(project.projectId)!.rules.push(rule);
+        });
+      } else {
+        teamRules.unassignedRules.push(rule);
+      }
+    });
+  });
+
+  teamMap.forEach((teamRules, teamId) => {
+    teamRules.projects = Array.from(teamProjectMaps.get(teamId)!.values());
+  });
+
+  return { teamRules: Array.from(teamMap.values()), unassignedToTeam };
+};
+
+/**
+ * Prepares one bucket's rules for the shared RuleRow tree.
+ * Ids are prefixed with their bucket since the same rule can live in multiple teams or projects
+ */
+export const scopeRulesToBucket = (bucketId: string, rules: Rule[]): { rows: Rule[]; rootIds: string[] } => {
+  const idsInBucket = new Set(rules.map((rule) => rule.ruleId));
+  const bucketPrefixId = (ruleId: string) => `${bucketId}::${ruleId}`;
+
+  // add bucket prefix id to each rule, filtering out subRuleIds that are not in this bucket
+  const rows: Rule[] = rules.map((rule) => ({
+    ...rule,
+    ruleId: bucketPrefixId(rule.ruleId),
+    subRuleIds: rule.subRuleIds.filter((id) => idsInBucket.has(id)).map(bucketPrefixId)
+  }));
+
+  // find top level rules in this bucket where parentRule is either not set or not in this bucket
+  const rootIds = rules
+    .filter((rule) => !rule.parentRule || !idsInBucket.has(rule.parentRule.ruleId))
+    .map((rule) => bucketPrefixId(rule.ruleId));
+
+  return { rows, rootIds };
+};
+
+/**
+ * Builds a structural header row (team, project, or "unassigned" section), not a real rule
+ */
+export const makeSectionRow = (ruleId: string, ruleCode: string, subRuleIds: string[]): Rule => ({
+  ruleId,
+  ruleCode,
+  ruleContent: '',
+  imageFileIds: [],
+  parentRule: undefined,
+  subRuleIds,
+  referencedRules: [],
+  isComplete: false
+});
